requirements.txt
apparmor==3.0.3
appdirs==1.4.4
application-utility==1.3.2
apsw==3.36.0.post1
asgiref==3.4.1
asn1crypto==1.4.0
Beaker==1.11.0
beautifulsoup4==4.9.3
btrfsutil==5.13.1
CacheControl==0.12.6
cchardet==2.1.7
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
css-parser==1.0.6
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
distlib==0.3.2
distro==1.5.0
dnspython==1.16.0
docopt==0.6.2
fastapi==0.68.0
feedparser==5.2.1
gufw==21.4.0
h11==0.12.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.1
httptools==0.2.0
idna==3.2
ifaddr==0.1.7
jeepney==0.7.1
keyutils==0.6
kiwisolver==1.3.1
lensfun==0.3.95
LibAppArmor==3.0.3
libtorrent===1.2.14-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.14-bindings-python
louis==3.18.0
lxml==4.6.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mechanize==0.4.5
more-itertools==8.8.0
msgpack==1.0.2
netifaces==0.11.0
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.20.3
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.5
pep517==0.11.0
Pillow==8.3.1
Pivy==0.6.6
ply==3.11
progress==1.5
psutil==5.8.0
pwquality==1.4.4
py7zr==0.11.3
pybind11==2.7.1
pycairo==1.20.0
pychm==0.8.6
pycparser==2.20
pycryptodome==3.10.1
pycups==2.0.1
pycurl==7.43.0.6
pydantic==1.8.2
Pygments==2.10.0
PyGObject==3.40.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
pyserial==3.5
PySide2==5.15.2
pysmbc==1.0.23
python-dateutil==2.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
regex==2021.7.6
reportlab==3.5.68
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
scipy==1.7.1
Shapely==1.7.1
shiboken2==5.15.2
sip==4.19.25
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
team==1.0
texttable==1.6.4
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
udiskie==2.3.3
ufw==0.36
unrardll==0.1.4
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
webencodings==0.5.1
websockets==9.1
zeroconf==0.29.0
